feat(admin): simplify CSV dataset + experiment upload
Datasets pane:
- Rename '+ Create Experiment' -> '+ Create Dataset and Experiment'.
- Add a 'sample format' CSV link under each create button (regular dataset
  columns; the experiment one has the extra 'expectation' column). Move the
  regular sample to csv.ts (SAMPLE_DATASET_CSV) so both panes share it.

Create Dataset + Experiment modal:
- Replace the long description with 'Upload dataset and expectation to create
  experiment. See the parent page for a link to a sample sheet.'
- Remove the in-modal sample-format link (now on the parent pane).
- Remove the 'Saved as <name>_dataset / <name>_experiment' helper line.

Docs (evaluation.md):
- Rename 'Quick path...' -> 'Quickstart: Load dataset and experiment in one file'
  and update the button-name references.
- Add a 'Manual path' section title so the step-by-step flow is clearly
  differentiated.
